# server/src/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Connection Manager API"
  version: "1.6.0"
  title: "Connection Manager API"
  contact:
    email: "ramiro@modusbox.com"
host: "localhost:3001"
basePath: "/api"
...
schemes:
- "http"
securityDefinitions:
  OAuth2:
    type: oauth2
    flow: "password"
     # The flow used by the OAuth2 security scheme. Valid values are "implicit", "password", "application" or "accessCode".
    tokenUrl: https://WSO2_IM_SERVERl:9443/oauth2/token
    scopes:
      pta: Portal Technical Administrator
      mta: Mojabox Technical Administrator
      everyone: All logged in users
security:
  - OAuth2: [ pta ] # most restrictive role
paths:
...
  /environments/{envId}/hub/cas:
    parameters:
    - name: "envId"
      in: "path"
      description: "ID of environment"
      required: true
      type: "string"
    post:
      tags:
      - "Hub-Certificate-Authority"
      summary: "Creates a new Hub CA"
      security:
        - OAuth2: [ pta ]
      description: |
        Creates a new Hub CA.
        
        * rootCertificate: a root certificate used by the Hub. Can be a self-signed certificate, or a globally trusted CA like Digicert. 
        * intermediateChain: list of intermediate certificates.
        
        The following combinations are possible
        
        * no rootCertificate and no intermediateChain: The Hub uses a trusted CA to sign its certificates.
        * rootCertificate and no intermediateChain: The Hub uses a self-signed CA to sign its certificates.
        * no rootCertificate and intermediateChain: The Hub uses a chain of intermediate certificates ultimately signed by a trusted CA to sign its certificates.
        * rootCertificate and intermediateChain: The Hub uses a chain of intermediate certificates ultimately signed by a self-signed CA to sign its certificates.
      operationId: "createHubCA"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CA info"
        required: true
        schema:
          $ref: "#/definitions/HubCAInfoInput"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HubCAInfo"
        400:
          description: "validation error"
      x-swagger-router-controller: "HubCAController"
    get:
      tags:
      - "Hub-Certificate-Authority"
      summary: "Returns the Hub CAs"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "getHubCAs"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HubCAInfo"
        400:
          description: "validation error"
      x-swagger-router-controller: "HubCAController"
  /environments/{envId}/hub/cas/{hubCAId}:
    parameters:
    - name: "envId"
      in: "path"
      description: "ID of environment"
      required: true
      type: "string"
    - name: "hubCAId"
      in: "path"
      description: "Hub CA Id"
      required: true
      type: "string"
    get:
      tags:
      - "Hub-Certificate-Authority"
      summary: "Returns the Hub CA with id = hubCAId"
      security:
        - OAuth2: [ mta, pta ]
      operationId: "getHubCA"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/HubCAInfo"
        400:
          description: "validation error"
      x-swagger-router-controller: "HubCAController"
    delete:
      tags:
      - "Hub-Certificate-Authority"
      summary: "Deletes the Hub CA with id = hubCAId"
      security:
        - OAuth2: [ pta ]
      operationId: "deleteHubCA"
      produces:
      - "application/json"
      responses:
        204:
          description: "successful operation"
        400:
          description: "validation error"
      x-swagger-router-controller: "HubCAController"
...
definitions:
...
  ValidationState:
    type: "string"
    enum:
      - "VALID"
      - "INVALID"
      - "NOT_AVAILABLE"
  HubCAInfoInput:
    type: "object"
    description: "Hub CA Input Body"
    required:
      - name
      - type
    properties:
      rootCertificate:
        type: "string"
        description: "PEM-encoded certificate"
      intermediateChain:
        type: "string"
        description: "PEM-encoded certificate"        
      name:
        type: "string"
        description: "Descriptive name. Must be unique in the environment"    
      type:
        type: "string"
        enum:
        - "EXTERNAL"
  HubCAInfo:
    type: "object"
    description: "Hub CA Input"
    allOf:
      - $ref: "#/definitions/HubCAInfoInput"
      - type: object
        properties:
          id:
            type: "integer"
            format: "int32"
          rootCertificateInfo:
            $ref: "#/definitions/CertInfo"
          intermediateChainInfo:
            type: "array"
            description: "List of CertInfo."
            items:
              $ref: "#/definitions/CertInfo"
          validations:
            $ref: "#/definitions/ValidationList"
          validationState:
            $ref: "#/definitions/ValidationState"
...
  CertInfo:
    type: "object"
    description: "Certificate contents and info"
    properties:
      subject:
        type: "object"
        properties:
          CN:
            type: "string"
            description: Common Name
          emailAddress:
            type: "string"
            description: email address
          O:
            type: "string"
            description: Organization
          OU:
            type: "string"
            description: Organizational Unit
          C:
            type: "string"
            description: Country
          ST:
            type: "string"
            description: State
          L:
            type: "string"
            description: Location
      issuer:
        type: "object"
        properties:
          CN:
            type: "string"
            description: Common Name
          emailAddress:
            type: "string"
            description: email address
          O:
            type: "string"
            description: Organization
          OU:
            type: "string"
            description: Organizational Unit
          C:
            type: "string"
            description: Country
          ST:
            type: "string"
            description: State
          L:
            type: "string"
            description: Location
      subjectAlternativeNames:
        type: "array"
        items:
          type: "string"
      serialNumber:
        type: "string"
        example: "17159976289428250189079625547205715159"
      notBefore:
        type: "string"
        description: "ISO date format"
        example: "2019-02-04T00:00:00Z"
      notAfter:
        type: "string"
        description: "ISO date format"
        example: "2020-02-12T12:00:00Z"
      signatureAlgorithm:
        type: "string"
        example: "SHA256WithRSA"
...
  ValidationCodes:
    type: "string"
    description:  "Identifies the type of validation"
    enum:
      - "CSR_CERT_SAME_PUBLIC_KEY"
      - "CSR_CERT_SAME_SUBJECT_INFO"
      - "CSR_CERT_PUBLIC_PRIVATE_KEY_MATCH"
      - "CERTIFICATE_SIGNED_BY_DFSP_CA"
      - "CERTIFICATE_VALIDITY"
      - "CERTIFICATE_USAGE_CLIENT"
      - "CSR_SIGNATURE_VALID"
      - "CSR_PUBLIC_KEY_LENGTH"
      - "CSR_SIGNATURE_ALGORITHM"
      - "CERTIFICATE_USAGE_SERVER"
      - "CERTIFICATE_CHAIN"
      - "CA_CERTIFICATE_USAGE"
      - "CERTIFICATE_VALID_OCSP"
      - "CERTIFICATE_VALID_CRL"
...
  Validation:
    type: "object"
    required:
      - validationCode
      - performed
      - result
      - message
    properties:
      validationCode:
        $ref: "#/definitions/ValidationCodes"
      performed:
        description: "Whether the validation has been performed or not"
        type: "boolean"
      result:
        type: "string"
        enum:
          - "VALID"
          - "INVALID"
          - "NOT_AVAILABLE"
      message:
        description: "Textual description of the validation result"
        type: "string"
      messageTemplate:
        description: "Textual description of the validation result, using the JavaScript template literal format"
        type: "string"
      data:
        description: "validation-specific data. Could be used by the UI to show more detail to the user"
        type: "object"
      details:
        description: "Command output or some other details about the validation"
        type: "string"
  ValidationList:
    type: "array"
    items:
      $ref: "#/definitions/Validation"
...